added history delete api
